+++ docs/.vuepress/components/Code.vue
@@ -7,7 +7,10 @@
 		/>
 		<slot v-else />
 
-		<div class="v-markup__copy">
+		<div
+			v-if="isClient"
+			class="v-markup__copy"
+		>
 			<VSlideXTransition>
 				<span
 					v-show="copied"
@@ -33,10 +36,12 @@
 </template>
 
 <script>
+	import Prism from 'vue-prism-component';
+
 	export default {
 		name: 'Code',
 		components: {
-			Prism: () => import('vue-prism-component')
+			Prism
 		},
 		data() {
 			return {
@@ -71,23 +76,25 @@
 	};
 </script>
 
-<style scoped>
+<style lang="scss" scoped>
 	.markup {
 		position: relative;
-	}
 
-	.markup >>> .v-markup__copy {
-		top: 0;
-		right: 0;
-		z-index: 1;
-		opacity: 0;
-		cursor: pointer;
-		transition: .25s;
-		position: absolute;
-	}
+		/deep/ .v-markup__copy {
+			top: 0;
+			right: 0;
+			z-index: 1;
+			opacity: 0;
+			cursor: pointer;
+			transition: .25s;
+			position: absolute;
+		}
 
-	.markup:hover >>> .v-markup__copy,
-	.markup:focus >>> .v-markup__copy {
-		opacity: 1;
+		&:hover,
+		&:focus {
+			/deep/ .v-markup__copy {
+				opacity: 1;
+			}
+		}
 	}
 </style>