+++ docs/.vuepress/components/Toolbar.vue
@@ -8,8 +8,10 @@
 	>
 		<VBtn
 			icon
-			class="mr-3 menu-btn"
-			:aria-label="$t('menu')"
+			class="pa-0 mr-3 menu-btn"
+			:to="isClient ? null : '/sitemap/'"
+			:aria-label="t('menu')"
+			:class="{ 'is-not-client': !isClient }"
 			@click="$emit('drawer:update')"
 		>
 			<SvgIcon icon="menu" />
@@ -17,7 +19,7 @@
 
 		<VBtn
 			flat
-			:to="$t('home')"
+			:to="t('home')"
 			active-class="active"
 			class="px-3 ma-0 text-none"
 		>
@@ -27,19 +29,20 @@
 			/>
 
 			<h1
-				v-show="!$vuetify.breakpoint.xsOnly"
+				v-show="!isClient || !$vuetify.breakpoint.xsOnly"
 				class="ml-2 mr-2 font-weight-regular"
 			>
-				{{ $t('name') }}
+				{{ t('name') }}
 			</h1>
 		</VBtn>
 
 		<VSpacer />
 
 		<VTextField
+			v-show="isClient"
 			solo
 			light
-			:label="$t('toolbar.search')"
+			:label="t('toolbar.search')"
 			single-line
 			hide-details
 			id="algolia-search-box"
@@ -48,6 +51,7 @@
 		/>
 
 		<VMenu
+			v-show="isClient"
 			left
 			bottom
 			offset-y
@@ -58,7 +62,7 @@
 				class="px-3 ma-0 text-none"
 			>
 				<template v-if="!$vuetify.breakpoint.xsOnly">
-					{{ $t('toolbar.ecosystem') }}
+					{{ t('toolbar.ecosystem') }}
 					<SvgIcon
 						size=".7em"
 						icon="down-arrow"
@@ -75,7 +79,7 @@
 
 			<VList>
 				<VListTile
-					v-for="item in $t('toolbar.items')"
+					v-for="item in t('toolbar.items')"
 					:key="item.title"
 					:href="item.href"
 					target="_blank"
@@ -120,20 +124,20 @@
 			menu-left
 			menu-bottom
 			:outline="false"
-			:value="$i18n.locale"
-			@change="updateLang"
+			:disabled="!isClient"
+			:value="currentLang"
 			:display-text-btn="false"
 			:display-arrow="false"
 			:available-languages="languages"
 			class="ma-0 text-none lang-btn"
 			flags-url="https://res.cloudinary.com/deraw/image/upload/v1547044454/"
+			@change="updateLang"
 		/>
 	</VToolbar>
 </template>
 
 <script>
 	import { version } from '../../../package.json';
-	import { loadLanguageAsync } from '../i18n';
 
 	export default {
 		name: 'Toolbar',
@@ -146,8 +150,6 @@
 		},
 		methods: {
 			updateLang(lang) {
-				loadLanguageAsync(lang);
-
 				if (this.initiated) {
 					this.$router.push(`/${lang}/`);
 				} else {
@@ -169,19 +171,23 @@
 						inputSelector: '#algolia-search-box'
 					});
 				}
-			}).catch((e) => console.log(e));
+			}).catch((e) => console.error(e));
 		}
 	}
 </script>
 
-<style scoped>
+<style lang="scss" scoped>
 	.v-toolbar {
-		z-index: 5 !important;
+		z-index: 7 !important;
+
+		/deep/ .v-toolbar__content {
+			height: 64px !important;
+		}
 	}
 
 	.v-toolbar .v-btn:not(.v-btn--icon),
 	.v-menu,
-	.v-menu >>> .v-menu__activator {
+	.v-menu /deep/ .v-menu__activator {
 		height: 100%;
 	}
 
@@ -189,21 +195,31 @@
 		flex: none;
 		height: 38px;
 		width: 250px;
+
+		/deep/ .v-input__slot {
+			border-radius: 50px !important;
+		}
+
+		/deep/ .v-input__control {
+			min-height: 38px !important;
+		}
 	}
 
-	.v-input.search >>> .v-input__slot {
-		border-radius: 50px !important;
+	.lang-btn {
+		/deep/ .v-btn {
+			margin: 0;
+			height: 100%;
+			padding: 20px;
+			min-width: 50px;
+		}
+
+		/deep/ .v-menu__activator--disabled {
+			cursor: default;
+		}
 	}
 
-	.v-input.search >>> .v-input__control {
-		min-height: 38px !important;
-	}
-
-	.lang-btn >>> .v-btn {
-		margin: 0;
-		height: 100%;
-		padding: 20px;
-		min-width: 50px;
+	.menu-btn.is-not-client {
+		display: none;
 	}
 
 	@media only screen and (max-width: 1000px) {
@@ -213,24 +229,30 @@
 		}
 
 		.v-toolbar .v-btn,
-		.lang-btn >>> .v-btn {
+		.lang-btn /deep/ .v-btn {
 			min-width: 0 !important;
 			padding: 0 16px !important;
 		}
 
-		.menu-btn {
+		.v-toolbar .menu-btn {
 			margin: 0 !important;
+			padding: 0 !important;
+			margin-right: 15px !important;
+		}
+
+		.menu-btn.is-not-client {
+			display: flex;
 		}
 	}
 
-	@media only screen and (max-width: 400px) {
+	@media only screen and (max-width: 500px) {
 		.v-toolbar .v-btn,
-		.lang-btn >>> .v-btn {
+		.lang-btn /deep/ .v-btn {
 			padding: 0 10px !important;
 		}
 	}
 
-	@media only screen and (max-width: 360px) {
+	@media only screen and (max-width: 400px) {
 		.v-input.search {
 			display: none;
 		}