+++ docs/.vuepress/enhanceApp.js
@@ -1,9 +1,14 @@
+import 'babel-polyfill';
 import Vuetify from 'vuetify';
-import i18n from './i18n';
+import VueDot from '@cnamts/vue-dot';
+import SSRMixin from './mixins/ssr';
 import { default as theme } from './theme.json';
 
 import 'prismjs';
 import 'prismjs/themes/prism-tomorrow.css';
+
+import './theme.css';
+import './vuetify.css';
 
 export default ({
 	Vue, // the version of Vue being used in the VuePress app
@@ -11,15 +16,19 @@ export default ({
 	router, // the router instance for the app
 	siteData // site metadata
 }) => {
-	Vue.use(Vuetify);
+	Vue.use(Vuetify, {
+		options: {
+			minifyTheme: (css) => {
+				return process.env.NODE_ENV === 'production'
+				? css.replace(/[\s|\r\n|\r|\n]/g, '')
+				: css
+			}
+		}
+	});
 
-	if (typeof document !== 'undefined') {
-		const VueDot = require('@cnamts/vue-dot');
+	Vue.use(VueDot, {
+		theme
+	});
 
-		Vue.use(VueDot, {
-			theme
-		});
-	}
-
-	options.i18n = i18n;
+	Vue.mixin(SSRMixin);
 };